Move things into parser package
